Más arreglos de bugs

+++ LaboratorioSaludTotal_Presentacion/src/main/java/Frames/RegistroParametroFORM.java
@@ -148,7 +148,7 @@ public class RegistroParametroFORM extends javax.swing.JFrame {
                     break;
                 case "Rango Inicial":
                     try {
-                        Double rangoInicial = Double.valueOf(entrada);
+                        Float rangoInicial = Float.valueOf(entrada);
                         if (rango.getRangoInicial() != null && rango.getRangoInicial().equals(rangoInicial)) {
                             resultados.add(rango);
                         }
@@ -158,7 +158,7 @@ public class RegistroParametroFORM extends javax.swing.JFrame {
                     break;
                 case "Rango Final":
                     try {
-                        Double rangoFinal = Double.valueOf(entrada);
+                        Float rangoFinal = Float.valueOf(entrada);
 
                         if (rango.getRangoFinal() != null && rango.getRangoFinal().equals(rangoFinal)) {
                             resultados.add(rango);
